Convert storage package and fix compilation issue.

+++ azure-mgmt-storage/src/main/java/com/azure/management/storage/PolicyRule.java
@@ -72,22 +72,22 @@ public interface PolicyRule extends
     /**
      * @return the number of days after a filtered base blob is last modified when the tier to cool action is enacted
      */
-    Double daysAfterBaseBlobModificationUntilCooling();
+    Integer daysAfterBaseBlobModificationUntilCooling();
 
     /**
      * @return the number of days after a filtered base blob is last modified when the tier to archive action is enacted
      */
-    Double daysAfterBaseBlobModificationUntilArchiving();
+    Integer daysAfterBaseBlobModificationUntilArchiving();
 
     /**
      * @return the number of days after a filtered base blob is last modified when the delete action is enacted
      */
-    Double daysAfterBaseBlobModificationUntilDeleting();
+    Integer daysAfterBaseBlobModificationUntilDeleting();
 
     /**
      * @return the number of days after a filtered snapshot is created when the delete action is enacted
      */
-    Double daysAfterSnapShotCreationUntilDeleting();
+    Integer daysAfterSnapShotCreationUntilDeleting();
 
     /**
      * Container interface for all of the definitions related to a rule in a management policy.
@@ -193,7 +193,7 @@ public interface PolicyRule extends
              * @param daysAfterBaseBlobModificationUntilCooling the number of days after a base blob is last modified until it is cooled.
              * @return the next stage of the management policy rule definition.
              */
-            WithPolicyRuleAttachable withTierToCoolActionOnBaseBlob(double daysAfterBaseBlobModificationUntilCooling);
+            WithPolicyRuleAttachable withTierToCoolActionOnBaseBlob(int daysAfterBaseBlobModificationUntilCooling);
 
             /**
              * The function that specifies a tier to archive action on the selected base blobs.
@@ -201,7 +201,7 @@ public interface PolicyRule extends
              * @param daysAfterBaseBlobModificationUntilArchiving the number of days after a base blob is last modified until it is archived.
              * @return the next stage of the management policy rule definition.
              */
-            WithPolicyRuleAttachable withTierToArchiveActionOnBaseBlob(double daysAfterBaseBlobModificationUntilArchiving);
+            WithPolicyRuleAttachable withTierToArchiveActionOnBaseBlob(int daysAfterBaseBlobModificationUntilArchiving);
 
             /**
              * The function that specifies a delete action on the selected base blobs.
@@ -209,7 +209,7 @@ public interface PolicyRule extends
              * @param daysAfterBaseBlobModificationUntilDeleting the number of days after a base blob is last modified until it is deleted.
              * @return the next stage of the management policy rule definition.
              */
-            WithPolicyRuleAttachable withDeleteActionOnBaseBlob(double daysAfterBaseBlobModificationUntilDeleting);
+            WithPolicyRuleAttachable withDeleteActionOnBaseBlob(int daysAfterBaseBlobModificationUntilDeleting);
 
             /**
              * The function that specifies a delete action on the selected snapshots.
@@ -217,7 +217,7 @@ public interface PolicyRule extends
              * @param daysAfterSnapShotCreationUntilDeleting the number of days after a snapshot is created until it is deleted.
              * @return the next stage of the management policy rule definition
              */
-            WithPolicyRuleAttachable withDeleteActionOnSnapShot(double daysAfterSnapShotCreationUntilDeleting);
+            WithPolicyRuleAttachable withDeleteActionOnSnapShot(int daysAfterSnapShotCreationUntilDeleting);
 
             /**
              * The function that specifies all of the actions to apply to selected base blobs.
@@ -318,7 +318,7 @@ public interface PolicyRule extends
              * @param daysAfterBaseBlobModificationUntilCooling the number of days after a base blob is last modified until it is cooled.
              * @return the next stage of the management policy rule update.
              */
-            Update withTierToCoolActionOnBaseBlob(double daysAfterBaseBlobModificationUntilCooling);
+            Update withTierToCoolActionOnBaseBlob(int daysAfterBaseBlobModificationUntilCooling);
 
             /**
              * The function that specifies a tier to archive action on the selected base blobs.
@@ -326,7 +326,7 @@ public interface PolicyRule extends
              * @param daysAfterBaseBlobModificationUntilArchiving the number of days after a base blob is last modified until it is archived.
              * @return the next stage of the management policy rule update.
              */
-            Update withTierToArchiveActionOnBaseBlob(double daysAfterBaseBlobModificationUntilArchiving);
+            Update withTierToArchiveActionOnBaseBlob(int daysAfterBaseBlobModificationUntilArchiving);
 
             /**
              * The function that specifies a delete action on the selected base blobs.
@@ -334,7 +334,7 @@ public interface PolicyRule extends
              * @param daysAfterBaseBlobModificationUntilDeleting the number of days after a base blob is last modified until it is deleted.
              * @return the next stage of the management policy rule update.
              */
-            Update withDeleteActionOnBaseBlob(double daysAfterBaseBlobModificationUntilDeleting);
+            Update withDeleteActionOnBaseBlob(int daysAfterBaseBlobModificationUntilDeleting);
 
             /**
              * The function that specifies a delete action on the selected snapshots.
@@ -342,7 +342,7 @@ public interface PolicyRule extends
              * @param daysAfterSnapShotCreationUntilDeleting the number of days after a snapshot is created until it is deleted.
              * @return the next stage of the management policy rule update
              */
-            Update withDeleteActionOnSnapShot(double daysAfterSnapShotCreationUntilDeleting);
+            Update withDeleteActionOnSnapShot(int daysAfterSnapShotCreationUntilDeleting);
 
             /**
              * The function that specifies all of the actions to apply to selected base blobs.
